Get past content type error
